# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Abuzar Next API
  version: 0.1.0
  description: Tenant-scoped central API for online and synchronized branch operations.
servers:
  - url: https://api.example.invalid
paths:
...
  /v1/auth/login:
    post:
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Session created; the response sets an HTTP-only cookie.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthenticatedSession' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /v1/tenants:
    get:
      operationId: listTenants
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Authenticated tenant
          content:
            application/json:
              schema:
                type: object
                required: [tenants]
                properties:
                  tenants: { type: array, items: { $ref: '#/components/schemas/Tenant' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/branches:
    get:
      operationId: listBranches
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Branches allowed by the authenticated operator
          content:
            application/json:
              schema:
                type: object
                required: [branches]
                properties:
                  branches: { type: array, items: { $ref: '#/components/schemas/Branch' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/counters:
    get:
      operationId: listCounters
      security: [{ sessionCookie: [] }]
      parameters:
        - name: branchId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Counters allowed by the authenticated operator
          content:
            application/json:
              schema:
                type: object
                required: [counters]
                properties:
                  counters: { type: array, items: { $ref: '#/components/schemas/Counter' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/access:
    get:
      operationId: currentAccess
      security: [{ sessionCookie: [] }]
      description: Effective normalized permissions plus imported legacy rights and scope rows. Unmapped right codes are reported as exceptions and are not claimed as exact command mappings.
      responses:
        '200':
          description: Effective access for the authenticated operator
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccessResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v1/roles:
    get:
      operationId: listRoles
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Tenant groups and their permission sets
          content:
            application/json:
              schema:
                type: object
                required: [roles]
                properties:
                  roles: { type: array, items: { $ref: '#/components/schemas/Role' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createRole
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoleUpsert' }
      responses:
        '201':
          description: Created tenant group
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Role' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/roles/{id}:
    patch:
      operationId: updateRole
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoleUpsert' }
      responses:
        '200':
          description: Updated tenant group
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Role' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/roles/{id}/rights:
    get:
      operationId: roleRights
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Imported rights and scopes for a tenant group
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoleRights' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateRoleRights
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoleRightsUpdate' }
      responses:
        '200':
          description: Updated imported rights and scopes
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoleRights' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/operators:
    get:
      operationId: listOperators
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Operators visible in the authenticated operational scope
          content:
            application/json:
              schema:
                type: object
                required: [operators]
                properties:
                  operators: { type: array, items: { $ref: '#/components/schemas/Operator' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/operators/{id}:
    patch:
      operationId: updateOperator
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OperatorUpdate' }
      responses:
        '200':
          description: Updated tenant-scoped operator
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operator' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/items/lookup:
    get:
      operationId: lookupItems
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: name, in: query, schema: { type: string } }
        - { name: alias, in: query, schema: { type: string } }
        - { name: barcode, in: query, schema: { type: string } }
        - { name: legacyId, in: query, schema: { type: string } }
      responses:
        '200':
          description: Tenant-scoped item matches by name, alias, barcode, or legacy identifier
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/ItemLookupResult' } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/inventory/availability:
    get:
      operationId: getInventoryAvailability
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: itemLegacyId, in: query, required: true, schema: { type: string } }
        - { name: godownId, in: query, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Active, non-expired stock batches for the selected tenant branch and godown
          content:
            application/json:
              schema:
                type: object
                required: [tenantId, branchId, itemLegacyId, godownId, batches]
                properties:
                  tenantId: { type: string, format: uuid }
                  branchId: { type: string, format: uuid }
                  itemLegacyId: { type: string }
                  godownId: { type: string, format: uuid }
                  batches:
                    type: array
                    items:
                      type: object
                      required: [batchId, batchNumber, quantity, unitCost]
                      properties:
                        batchId: { type: string, format: uuid }
                        batchNumber: { type: string }
                        expiryDate: { type: string, format: date }
                        quantity: { type: string }
                        unitCost: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/master/{kind}:
    get:
      operationId: listCanonicalMaster
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: kind, in: path, required: true, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        '200':
          description: Tenant-scoped canonical master list
          content:
            application/json:
              schema:
                type: object
                required: [records]
                properties:
                  records: { type: array, items: { $ref: '#/components/schemas/MasterRecord' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createCanonicalMaster
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: kind, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MasterUpsert' } } }
      responses:
        '201':
          description: Created canonical master
          content: { application/json: { schema: { $ref: '#/components/schemas/MasterRecord' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/master/{kind}/{id}:
    get:
      operationId: getCanonicalMaster
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: kind, in: path, required: true, schema: { type: string } }
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Tenant-scoped canonical master detail
          content: { application/json: { schema: { $ref: '#/components/schemas/MasterRecord' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateCanonicalMaster
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: kind, in: path, required: true, schema: { type: string } }
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MasterUpsert' } } }
      responses:
        '200':
          description: Updated canonical master
          content: { application/json: { schema: { $ref: '#/components/schemas/MasterRecord' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      operationId: deleteCanonicalMaster
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: kind, in: path, required: true, schema: { type: string } }
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Canonical master deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/master/item/{id}/suppliers:
    get:
      operationId: listItemSuppliers
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Tenant-scoped ItemSuppliers grid
          content:
            application/json:
              schema:
                type: object
                required: [suppliers]
                properties:
                  suppliers: { type: array, items: { $ref: '#/components/schemas/ItemSupplier' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: replaceItemSuppliers
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [suppliers]
              properties:
                suppliers: { type: array, items: { $ref: '#/components/schemas/ItemSupplier' } }
      responses:
        '200': { description: Replaced ItemSuppliers grid }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/transactions/sales:
    post:
      operationId: createSale
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SyncEnvelope' }
      responses:
        '202':
          description: Sale accepted or deduplicated
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted: { type: boolean }
                  duplicate: { type: boolean }
                  eventId: { type: string, format: uuid }
                  aggregateId: { type: string, format: uuid }
        '409': { $ref: '#/components/responses/Conflict' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/transactions/returns:
    post:
      operationId: createReturn
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/SyncEnvelope' } }
      responses:
        '202':
          description: Return accepted or deduplicated
          content:
            application/json: { schema: { $ref: '#/components/schemas/TransactionResult' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/transactions/receiving:
    post:
      operationId: createReceiving
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/SyncEnvelope' } }
      responses:
        '202':
          description: Receiving accepted or deduplicated
          content:
            application/json: { schema: { $ref: '#/components/schemas/TransactionResult' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/transactions/inventory:
    post:
      operationId: createInventoryTransaction
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/SyncEnvelope' } }
      responses:
        '202':
          description: Inventory event accepted or deduplicated
          content:
            application/json: { schema: { $ref: '#/components/schemas/TransactionResult' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/tax-rates:
    get:
      operationId: listTaxRates
      security: [{ sessionCookie: [] }]
      parameters:
        - name: effectiveAt
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Effective-dated tax rates in the authenticated branch
          content:
            application/json:
              schema:
                type: object
                required: [rates]
                properties:
                  rates: { type: array, items: { $ref: '#/components/schemas/TaxRate' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createTaxRate
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/TaxRateUpsert' } }
      responses:
        '201': { description: Tax rate created, content: { application/json: { schema: { $ref: '#/components/schemas/TaxRate' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/tax-rates/{id}:
    patch:
      operationId: updateTaxRate
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/TaxRateUpsert' } }
      responses:
        '200': { description: Tax rate updated, content: { application/json: { schema: { $ref: '#/components/schemas/TaxRate' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteTaxRate
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Tax rate deleted }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/tax-assignments:
    get:
      operationId: listTaxAssignments
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: itemId, in: query, schema: { type: string, format: uuid } }
        - { name: partyId, in: query, schema: { type: string, format: uuid } }
        - { name: effectiveAt, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: Effective-dated item and party tax assignments
          content:
            application/json:
              schema:
                type: object
                required: [items, parties]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/TaxAssignment' } }
                  parties: { type: array, items: { $ref: '#/components/schemas/TaxAssignment' } }
  /v1/tax-assignments/items/{itemId}:
    put:
      operationId: replaceItemTaxAssignments
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/TaxAssignmentReplace' } }
      responses:
        '200': { description: Item tax assignments replaced }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/tax-assignments/parties/{partyId}:
    put:
      operationId: replacePartyTaxAssignments
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: partyId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/TaxAssignmentReplace' } }
      responses:
        '200': { description: Party tax assignments replaced }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/tax-assignments/apply-item-gst:
    post:
      operationId: applyItemGST
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/ApplyItemGST' } }
      responses:
        '200': { description: GST assignment applied to the selected or all active canonical items }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/documents/{kind}:
    post:
      operationId: commandBusinessDocument
      description: |
        Commands cash/credit sales, cash/credit sale returns, quotations, refused sales,
        and the Phase I purchase families:
        pack-purchase, loose-purchase, opening-purchase, purchase-return, and
        purchase-order. Sale returns require a posted source sale and restore its
        source batches before reversing finance projections. Save, post,
        save-and-post, and void are idempotent by
        commandId/idempotencyKey and use optimistic document revisions.
      security: [{ sessionCookie: [] }]
      parameters:
        - name: kind
          in: path
          required: true
          schema: { $ref: '#/components/schemas/DocumentKind' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DocumentCommand' }
      responses:
        '200':
          description: Document command accepted or replayed from an idempotent receipt
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentCommandResult' }
        '201':
          description: New draft document saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentCommandResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
...
  /v1/shifts/open:
    post:
      operationId: openShift
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount: { type: string }
      responses:
        '201': { description: Shift opened }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: abuzar_session
  responses:
    Unauthorized:
      description: Authentication required
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
...
    Unavailable:
      description: Database or identity store unavailable
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
...
  schemas:
    LoginRequest:
      type: object
      required: [username, password, tenantCode]
      properties:
        username: { type: string, minLength: 1 }
        password: { type: string, minLength: 1, writeOnly: true }
        tenantCode: { type: string, minLength: 1 }
        branchId: { type: string, format: uuid }
        counterId: { type: string, format: uuid }
...
    AuthenticatedSession:
      type: object
      required: [authenticated, context, expiresAt]
      properties:
        authenticated: { const: true }
        context: { $ref: '#/components/schemas/TenantContext' }
        expiresAt: { type: string, format: date-time }
    Tenant:
      type: object
      required: [id, code, legalName, active]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        legalName: { type: string }
        active: { type: boolean }
...